accurate score + toggle 2dmv line effect

+++ Assets/Scripts/Assembly-CSharp/Sekai/Core/Live/TechnicScoreLogic.cs
@@ -9,7 +9,7 @@ namespace Sekai.Core.Live
 		{
 		}
 
-		public int GetScoreValueForm(int type)
+		public float GetScoreValueForm(int type)
 		{
 			return BaseNoteScore;
 		}
